refactor: moved onboarding
