+++ src/features/wallet/routers/index.ts
@@ -1,4 +1,3 @@
-import "./onboarding";
 import "./payment";
 import "./wallets";
